requirements.txt
flask
opencv-python
numpy
Pillow
docx
datetime
pytesseract